llm generates report

+++ Backend/learn-service/src/main/java/org/ba/repositories/LearnerRepository.java
@@ -6,6 +6,6 @@ import io.quarkus.hibernate.orm.panache.PanacheRepositoryBase;
 import jakarta.enterprise.context.ApplicationScoped;
 
 @ApplicationScoped
-public class LearnerRepository implements PanacheRepositoryBase<LearnerEntity, Integer> {
+public class LearnerRepository implements PanacheRepositoryBase<LearnerEntity, Long> {
      
 }
